+++ src/main/webapp/employee/books.jsp
@@ -63,7 +63,7 @@
                                     <td><%=books.get(i).getCategories()%></td>
                                     <td><%=books.get(i).getPublishDate()%></td>
                                     <td><a href="update-book.jsp?ID=<%=books.get(i).getID()%>" class="btn btn-warning"><i class="ti-pencil"></i> Güncelle</a></td>
-                                    <td><a href="delete-book.jsp?ID=<%=books.get(i).getID()%>" class="btn btn-danger"><i class="icon-trash"></i> Sil</a></td>
+                                    <td><a href="delete-book.jsp?ID=<%=books.get(i).getID()%>" class="btn btn-danger" onclick="return confirm('Emin misin?')"><i class="icon-trash"></i> Sil</a></td>
                                 </tr>
                             <%        
                                 }
